+++ CHANGELOG.md
@@ -115,6 +115,7 @@ hi
 - Carnations now yield magenta dye instead of purple dye.
 - Added Frozen Short Grass, Frozen Tall Grass, Frozen Fern, and Frozen Large Fern blocks.
   - Shoutout to just_jose2006 for the textures!
+- Changed the Warm River's surface rules to likely reduce lag caused by Sand blocks falling.
 
 # Music
 - Added three new tracks by Willow/pictochats_.